refactor: extract litellm functions into _litellm.py backend module
Move all litellm completion functions, capability queries, streaming
helpers, and request transformers from lm.py into _litellm.py with
generic names (complete, acomplete, text_complete, etc.).

LM class body is now litellm-free — it imports generic function names
from _litellm and delegates to them. Old function names (litellm_completion,
alitellm_completion, etc.) are re-exported as aliases for backward compat.

3 files changed. Zero behavior change. 754 tests pass.
